1.修改表结构 增加审核字段 2.增加审核界面 3.前端一波优化

+++ src/main/java/com/hf/laf/web/FoundController.java
@@ -5,50 +5,46 @@ import com.hf.laf.configs.PathConfig;
 import com.hf.laf.dto.PaginationDto;
 import com.hf.laf.entity.FoundRegister;
 import com.hf.laf.service.FoundRegisterService;
-import com.sun.istack.internal.Nullable;
+import com.hf.laf.util.AdminTool;
+import com.hf.laf.util.FileUtil;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpRequest;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-import util.FileUtil;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.OutputStream;
+import java.util.Date;
 
 @Controller
 @RequestMapping("/found")
 public class FoundController {
 
     @Autowired
-    private PathConfig pathConfig;
-
-    @Autowired
     FoundRegisterService foundRegisterService;
+    @Autowired
+    private PathConfig pathConfig;
 
     @RequestMapping("/select")
     @ResponseBody
-    public Object select(Integer foundId) {
-        return foundRegisterService.selectFoundRegisterById(foundId);
+    public Object select(Integer foundId, HttpSession session) {
+        if (AdminTool.isAdmin(session)) {
+            return foundRegisterService.selectFoundRegisterById(foundId);
+        }
+        return foundRegisterService.selectFoundRegisterByIdNotAuth(foundId);
     }
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public Object add(FoundRegister record,@RequestParam("picInput") MultipartFile file) {
+    public Object add(FoundRegister record, @RequestParam("picInput") MultipartFile file) {
 
+        String fileName = String.valueOf(new Date().getTime());
 
-        String fileName = file.getOriginalFilename();
-        /*System.out.println("fileName-->" + fileName);
-        System.out.println("getContentType-->" + contentType);*/
+        if (!file.isEmpty()) {
 
-
-        if(!file.isEmpty()){
-
-            String filePath =pathConfig.getFoundPicPath();
+            String filePath = pathConfig.getFoundPicPath();
             try {
                 FileUtil.uploadFile(file.getBytes(), filePath, fileName);
             } catch (Exception e) {
@@ -56,51 +52,65 @@ public class FoundController {
             }
             record.setFoundPic(fileName);
         }
-        if (foundRegisterService.addRecord(record) == 1){
+        if (foundRegisterService.addRecord(record) == 1) {
             return "foundList";
-        }
-        else return "addError";
+        } else return "addError";
     }
 
-    @RequestMapping(value = "/delete",method = RequestMethod.POST)
+    @RequestMapping(value = "/delete", method = RequestMethod.GET)
     @ResponseBody
-    public Object delete(Integer foundId,HttpSession session) {
-        if (session.getAttribute("isadmin").equals("666") ){
-        return foundRegisterService.deleteRecord(foundId);
+    public Object delete(Integer foundId, HttpSession session) {
+        if (AdminTool.isAdmin(session)) {
+            return foundRegisterService.deleteRecord(foundId);
         }
         return "permissionError";
     }
 
-    @RequestMapping(value = "/update",method = RequestMethod.POST)
+    @RequestMapping(value = "/update", method = RequestMethod.POST)
     @ResponseBody
-    public Object update(FoundRegister record,HttpSession session) {
-        if (session.getAttribute("isadmin").equals("666") ){
-        return foundRegisterService.updateRecord(record);
+    public Object update(FoundRegister record, HttpSession session) {
+        if (AdminTool.isAdmin(session)) {
+            return foundRegisterService.updateRecord(record);
+        }
+        return "permissionError";
     }
-        return  "permissionError";
-    }
 
-
-
-    @RequestMapping("/selectpagination")
+    @RequestMapping(value = "/setReview", method = RequestMethod.POST)
     @ResponseBody
-    public Object selectpagination(@RequestParam(defaultValue = "10") int limit, @RequestParam(defaultValue = "0")  int offset) {
+    public Object update(Integer foundId, Integer hadReview, HttpSession session) {
+        if (AdminTool.isAdmin(session)) {
+            return foundRegisterService.setReview(hadReview, foundId);
+        }
+        return "permissionError";
+    }
+
+
+    @RequestMapping("/selectPagination")
+    @ResponseBody
+    public Object selectPagination(@RequestParam(defaultValue = "10") int limit, @RequestParam(defaultValue = "0") int offset, HttpSession session) {
         PaginationDto pag = new PaginationDto();
-        pag.setTotal(foundRegisterService.count());
-        pag.setRows(foundRegisterService.selectPaginationRecord(limit,offset));
+
+        if(AdminTool.isAdmin(session)){
+            pag.setTotal(foundRegisterService.count());
+            pag.setRows(foundRegisterService.selectPaginationRecord(limit, offset));
+        }else{
+            pag.setTotal(foundRegisterService.countReviewed());
+            pag.setRows(foundRegisterService.selectReviewedPaginationRecord(limit, offset));
+        }
+
 
         return pag;
 
     }
 
     @RequestMapping("/pics/{fileName:[a-zA-Z0-9\\\\.]+}")
-    public void getPic(@PathVariable String fileName, HttpServletResponse response){
+    public void getPic(@PathVariable String fileName, HttpServletResponse response) {
         File file = new File(pathConfig.getFoundPicPath());
         try {
-            FileInputStream fileInputStream = new FileInputStream(file + File.separator +  fileName);
+            FileInputStream fileInputStream = new FileInputStream(file + File.separator + fileName);
             OutputStream out = response.getOutputStream();
 
-            FileUtil.streamCopy(out,fileInputStream);
+            FileUtil.streamCopy(out, fileInputStream);
 
             fileInputStream.close();
         } catch (Exception e) {
@@ -110,6 +120,7 @@ public class FoundController {
 
     /**
      * 拾物列表页
+     *
      * @return 列表页文件名
      */
     @RequestMapping("/list")
@@ -121,10 +132,11 @@ public class FoundController {
 
     /**
      * 新拾物发表页
+     *
      * @return 发表页文件名
      */
-    @RequestMapping(value = "/add",method = RequestMethod.GET)
-    public String newItem(){
+    @RequestMapping(value = "/add", method = RequestMethod.GET)
+    public String newItem() {
         return "newFoundItem";
     }
 
@@ -132,7 +144,7 @@ public class FoundController {
      *
      */
     @RequestMapping(value = "/detail")
-    public String detailPage(){
+    public String detailPage() {
         return "foundDetail";
     }
